+++ web/components/Forms/PwChangeForm/styled.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-const InputForm = styled.form`
+const InputForm = styled.div`
   width: 300px;
   display: flex;
   flex-direction: column;
@@ -51,7 +51,7 @@ const ErrorText = styled.span`
 `;
 
 const Title = styled.span`
-  font-size: 1.8rem;
+  font-size: 1.3rem;
   font-weight: bold;
   display: flex;
   align-items: center;
